@@ -5,7 +5,7 @@
   },
   "name": "mSupplyMobile",
   "//": "version must be in the format ${majorNumber}.${minorNumber}.${patchNumber}-rc${releaseCandidateNumber}",
-  "version": "8.2.3",
+  "version": "8.3.0",
   "private": false,
   "license": "MIT",
   "description": "Mobile app for use with the mSupply medical inventory control software",
@@ -49,7 +49,7 @@
     ]
   },
   "dependencies": {
-    "@bugsnag/react-native": "^7.9.6",
+    "@bugsnag/react-native": "^7.11.0",
     "@react-native-async-storage/async-storage": "^1.15.4",
     "@react-native-community/checkbox": "^0.5.8",
     "@react-native-community/datetimepicker": "^3.4.7",
@@ -67,6 +67,7 @@
     "buffer": "^6.0.3",
     "currency.js": "^2.0.3",
     "jest": "^26.6.3",
+    "jsc-android": "^250230.2.1",
     "json-pointer": "^0.6.1",
     "json2csv": "^5.0.6",
     "lodash.merge": "^4.6.2",
@@ -121,9 +122,13 @@
     "@babel/core": "^7.14.0",
     "@babel/eslint-parser": "^7.14.2",
     "@babel/runtime": "^7.14.0",
+    "@types/react": "^17.0.15",
+    "@types/react-native": "^0.64.12",
+    "@typescript-eslint/eslint-plugin": "^4.29.0",
+    "@typescript-eslint/parser": "^4.29.0",
     "babel-preset-react-native": "^4.0.1",
-    "eslint": "^7.25",
-    "eslint-config-airbnb": "^18.0.1",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.4.0",
@@ -131,6 +136,8 @@
     "husky": "^6.0.0",
     "lint-staged": "^10.5.4",
     "metro-react-native-babel-preset": "^0.66.0",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "react-native-typescript-transformer": "^1.2.13",
+    "typescript": "^4.3.5"
   }
 }
