@@ -5,7 +5,9 @@
 
 import DeviceInfo from 'react-native-device-info';
 
-import { authenticateAsync, AUTH_ERROR_CODES, hashPassword } from 'sussol-utilities';
+import { hashPassword } from 'sussol-utilities';
+import { AUTH_ERROR_CODES, authenticateAsync } from './Helpers';
+import LoggerService from '../utilities/logging';
 
 import { SETTINGS_KEYS } from '../settings';
 
@@ -16,6 +18,8 @@ const { CONNECTION_FAILURE, INVALID_PASSWORD } = AUTH_ERROR_CODES;
 const AUTH_ENDPOINT = '/sync/v3/user';
 
 const CONNECTION_TIMEOUT_PERIOD = 10 * 1000; // 10 second timeout for authenticating connection.
+
+const logger = LoggerService.createLogger('Authentication');
 
 export class UserAuthenticator {
   constructor(database, settings) {
@@ -64,20 +68,25 @@ export class UserAuthenticator {
         authenticateAsync(authURL, username, passwordHash, { ...this.extraHeaders }),
         createConnectionTimeoutPromise(),
       ]);
+      if (userJson && userJson.error) {
+        throw new Error(userJson.error);
+      }
       if (!userJson || !userJson.UserID) {
         throw new Error('Unexpected response from server');
-      } else {
-        // Success, save user to database.
-        this.database.write(() => {
-          user = this.database.update('User', {
-            id: userJson.UserID,
-            username,
-            passwordHash,
-            isAdmin: userJson.isAdmin || false,
-          });
-        });
       }
+
+      // Success, save user to database.
+      this.database.write(() => {
+        user = this.database.update('User', {
+          id: userJson.UserID,
+          username,
+          passwordHash,
+          isAdmin: userJson.isAdmin || false,
+        });
+      });
     } catch (error) {
+      logger.error(error);
+
       // If there was an error with connection, check against locally cached credentials.
       if (error.message === CONNECTION_FAILURE && user) {
         if (user.username === username && user.passwordHash === passwordHash) {
@@ -122,6 +131,7 @@ const createConnectionTimeoutPromise = () =>
   new Promise((resolve, reject) => {
     const id = setTimeout(() => {
       clearTimeout(id);
+      logger.error('UserAuthenticator: connection timeout');
       reject(new Error(CONNECTION_FAILURE));
     }, CONNECTION_TIMEOUT_PERIOD);
   });
