@@ -12,8 +12,7 @@ import { createRecord } from '../database/utilities';
 import { navStrings } from '../localization';
 import { ROUTES } from './constants';
 import { RootNavigator } from './RootNavigator';
-import { PrescriptionActions } from '../actions/PrescriptionActions';
-import { FinaliseActions } from '../actions/FinaliseActions';
+import { PrescriptionActions, FinaliseActions } from '../actions/index';
 import { PREFERENCE_KEYS } from '../database/utilities/constants';
 import { SensorActions } from '../actions/Entities/SensorActions';
 import { TemperatureBreachConfigActions } from '../actions/Entities/TemperatureBreachConfigActions';
