@@ -5,6 +5,8 @@ import { selectSpecificEntityState } from './index';
 import { DATE_FORMAT } from '../../utilities/constants';
 import { PREFERENCE_KEYS } from '../../database/utilities/preferenceConstants';
 import { MILLISECONDS_PER_DAY } from '../../database/utilities/constants';
+import { validateJsonSchemaData } from '../../utilities';
+import { convertMobileDateToISO } from '../../utilities/formatters';
 
 export const selectEditingNameId = state => {
   const NameState = selectSpecificEntityState(state, 'name');
@@ -77,16 +79,49 @@ export const selectCanEditPatient = state => {
   return UIDatabase.getPreference(PREFERENCE_KEYS.CAN_EDIT_PATIENTS_FROM_ANY_STORE) || isEditable;
 };
 
-export const selectVaccinePatientHistory = state => {
-  const patientId = selectEditingNameId(state) ?? '';
+const jsonSchema = {
+  type: 'object',
+  properties: {
+    refused: {
+      type: 'boolean',
+      enum: [false],
+    },
+    vaccinator: {
+      type: 'string',
+    },
+    itemName: {
+      type: 'string',
+    },
+    itemCode: {
+      type: 'string',
+    },
+    vaccineDate: {
+      type: 'string',
+    },
+  },
+};
 
-  const inQuery = UIDatabase.objects('Transaction')
-    .filtered('otherParty.id == $0', patientId)
-    .map(({ id }) => `transaction.id == "${id}"`)
-    .join(' OR ');
-  const baseQueryString = 'type != "cash_in" AND type != "cash_out"';
-  const fullQuery = `(${inQuery}) AND ${baseQueryString} AND itemBatch.item.isVaccine == true`;
-  return inQuery ? UIDatabase.objects('TransactionBatch').filtered(fullQuery).slice() : [];
+export const selectVaccinePatientHistory = patient => {
+  const [vaccinationPatientEvent] = UIDatabase.objects('PatientEvent').filtered(
+    "code == 'vaccination'"
+  );
+  const { id: vaccinationPatientEventID } = vaccinationPatientEvent ?? {};
+
+  const nameNotes = patient?.nameNotes
+    ?.filter(
+      ({ patientEventID, data }) =>
+        patientEventID === vaccinationPatientEventID && validateJsonSchemaData(jsonSchema, data)
+    )
+    .map(({ id, data: vaccinationNameNotes }) => ({
+      ...vaccinationNameNotes,
+      id,
+      doses: 1, // Currently not possible to dispense more than 1 dose
+      totalQuantity: 1,
+      confirmDate: new Date(convertMobileDateToISO(vaccinationNameNotes.vaccineDate)),
+      prescriberOrVaccinator: vaccinationNameNotes.vaccinator,
+    }));
+
+  return nameNotes ?? [];
 };
 
 export const selectWasPatientVaccinatedWithinOneDay = state => {
