@@ -40,7 +40,7 @@ LoadingIndicator.propTypes = {
   loading: PropTypes.bool.isRequired,
 };
 
-const getColumnKey = (isVaccine, vaccineDispensingEnabled, canViewHistory) => {
+const getColumnKey = (isVaccine, canViewHistory) => {
   // Vaccine Dispensing History
   if (isVaccine) {
     return canViewHistory ? MODALS.VACCINE_HISTORY_LOOKUP : MODALS.VACCINE_HISTORY;
@@ -51,23 +51,20 @@ const getColumnKey = (isVaccine, vaccineDispensingEnabled, canViewHistory) => {
     return MODALS.PATIENT_HISTORY;
   }
 
-  return vaccineDispensingEnabled
-    ? MODALS.PATIENT_HISTORY_LOOKUP_WITH_VACCINES
-    : MODALS.PATIENT_HISTORY_LOOKUP;
+  return MODALS.PATIENT_HISTORY_LOOKUP_WITH_VACCINES;
 };
 
-export const PatientHistoryModal = ({
-  isVaccine,
-  patientId,
-  patientHistory,
-  sortKey,
-  vaccineDispensingEnabled,
-}) => {
+export const PatientHistoryModal = ({ isVaccine, patientId, patientHistory, sortKey }) => {
   const canViewHistory = UIDatabase.getPreference(PREFERENCE_KEYS.CAN_VIEW_ALL_PATIENTS_HISTORY);
-  const columns = React.useMemo(
-    () => getColumns(getColumnKey(isVaccine, vaccineDispensingEnabled, canViewHistory)),
-    []
+  const patientsSyncEverywhere = !UIDatabase.getPreference(
+    PREFERENCE_KEYS.NEW_PATIENTS_VISIBLE_THIS_STORE_ONLY
   );
+
+  // Remote fetch not required if patients sync everywhere is enabled and only fetching vaccines
+  const isRemoteFetchRequired =
+    (canViewHistory && !isVaccine) || (canViewHistory && !patientsSyncEverywhere && isVaccine);
+
+  const columns = React.useMemo(() => getColumns(getColumnKey(isVaccine, canViewHistory)), []);
   const [{ data, loading, error }, fetchOnline] = useLocalAndRemotePatientHistory({
     isVaccineDispensingModal: isVaccine,
     patientId,
@@ -75,7 +72,7 @@ export const PatientHistoryModal = ({
     sortKey,
   });
 
-  if (canViewHistory) {
+  if (isRemoteFetchRequired) {
     useEffect(fetchOnline, [patientId]);
     useEffect(
       () =>
@@ -103,13 +100,12 @@ const localStyles = {
 
 PatientHistoryModal.defaultProps = {
   isVaccine: false,
-  vaccineDispensingEnabled: false,
+  patientHistory: [],
 };
 
 PatientHistoryModal.propTypes = {
   isVaccine: PropTypes.bool,
   patientId: PropTypes.string.isRequired,
-  patientHistory: PropTypes.array.isRequired,
+  patientHistory: PropTypes.array,
   sortKey: PropTypes.string.isRequired,
-  vaccineDispensingEnabled: PropTypes.bool,
 };
